Refactor Awareness to Traits
Changed the Awareness folder to traits. Unfortunately 'trait'
which I think both of us would have preferred is a reserved
word and therefore cannot be used as part of a namespace.
