+++ src/Context/Traits/WidgetAwareContextTrait.php
@@ -1,5 +1,5 @@
 <?php
-namespace PaulGibbs\WordpressBehatExtension\Context\Awareness;
+namespace PaulGibbs\WordpressBehatExtension\Context\Traits;
 
 trait WidgetAwareContextTrait
 {
